Refactored packages, fixed API
