--- src/main/java/com/tvseries/TvSeries/TvShowRepository.java
@@ -1,9 +1,0 @@
-package com.tvseries.TvSeries;
-
-import com.tvseries.TvSeries.dto.TvShow;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-
-interface TvShowRepository extends JpaRepository<TvShow, Long> {
-
-}
