+++ src/main/java/com/tvseries/TvSeries/db/UserRepository.java
@@ -1,7 +1,6 @@
-package com.tvseries.TvSeries;
+package com.tvseries.TvSeries.db;
 
-import com.tvseries.TvSeries.dto.TvShow;
-import com.tvseries.TvSeries.dto.User;
+import com.tvseries.TvSeries.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository  extends JpaRepository<User, Long> {
