+++ src/main/java/com/tvseries/TvSeries/model/User.java
@@ -1,15 +1,8 @@
-package com.tvseries.TvSeries.dto;
-
-import org.hibernate.annotations.Fetch;
-import org.springframework.data.repository.cdi.Eager;
+package com.tvseries.TvSeries.model;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
-import javax.persistence.OneToOne;
-import javax.transaction.Transactional;
-import java.util.ArrayList;
-import java.util.List;
 
 @Entity
 public class User {
